refactor(agent): extract bootstrap-context helpers (#3250)
Behavior-preserving extraction of the pure fork / dropped-prompt-recovery
prompt-text cluster from agent.store.ts (#3250) into a co-located module
src/lib/agent/bootstrap-context.ts.

Moved (pure — zero `this.`, zero closure over store state or services):
- agentDisplayName, agentInitializationFailureMessage
- truncateBootstrapText, formatForkBootstrapMessage (module-private — no
  remaining store callers), FORK_BOOTSTRAP_MAX_MSG_CHARS
- buildForkBootstrapContext
- isSessionDeathMessage, isRecoverableDeadSessionSendFailure,
  filterDroppedPromptRecoveryMessages
- mergeRecoveryMessages, buildDroppedPromptRecoveryBootstrapContext

ActiveSession / AgentMessage are type-only imports (erased at runtime, no
cycle). Deliberately NOT moved: refreshClaudeMemoryMdBeforeSpawn (calls the
claude-memory render service) and buildDroppedPromptRecoverySnapshot (reads
SQLite via loadPersistedAgentHistory) — both stay in the store. Every call
site is unchanged; the store re-exports agentDisplayName so its two
component importers (ModelSelector, ThreadProviderSwitcher) and the public
surface stay identical. agent.store.ts drops from 9101 to 8918 lines.

Three tests pin the moved declarations by SOURCE TEXT; those assertions are
repointed to bootstrap-context.ts where the code now lives — same fact,
same body asserted, zero weakening. Assertions about the store's own
wiring (e.g. that it references agentInitializationFailureMessage) stay
pointed at agent.store.ts.

No behavior, IPC-contract, or persistence-schema change.

Validation: full unit suite green (2708 passed — identical counts to clean
main at the same base); Biome clean; all changed files type-clean under tsc.

Taariq Lewis, SerenAI, Paloma, and Volume at https://serendb.com
Email: hello@serendb.com
