+++ tests/unit/agent-spawn-guard.test.ts
@@ -59,11 +59,15 @@ describe("agent initialization failure copy", () => {
     resolve("src/stores/agent.store.ts"),
     "utf-8",
   );
+  const bootstrapContextSource = readFileSync(
+    resolve("src/lib/agent/bootstrap-context.ts"),
+    "utf-8",
+  );
 
   it("uses provider-specific fallback remediation instead of a Claude-only message", () => {
     expect(agentStoreSource).toContain("agentInitializationFailureMessage");
-    expect(agentStoreSource).toContain("Codex is installed and signed in");
-    expect(agentStoreSource).not.toContain(
+    expect(bootstrapContextSource).toContain("Codex is installed and signed in");
+    expect(bootstrapContextSource).not.toContain(
       "Agent session terminated before initialization completed. Check that Claude Code is installed and authenticated.",
     );
   });
